+++ tests/test_pixel_level_inputs.py
@@ -54,6 +54,19 @@ def test_basic_additive_image():
     assert image['layout']['uirevision']
 
 
+    mask_config = {"mask": np.zeros((600, 600, 3)).astype(np.uint8)}
+
+    image_mask = get_additive_image_with_masking(["DNA", "Nuclear"], data_selection="experiment0+++slide0+++acq0",
+                                            canvas_layers=upload_dict, mask_config=mask_config, mask_toggle=True,
+                                            mask_selection="mask", show_canvas_legend=True, mask_blending_level=1,
+                                            add_mask_boundary=True, legend_text='')
+    assert isinstance(image_mask, go.Figure)
+    assert image_mask['data'] is not None
+    assert image_mask['data'][0]['hovertemplate'] == 'x: %{x}<br>y: %{y}<br><extra></extra>'
+    assert image_mask['layout']['annotations'][0]['text'] == '<span style="color: white">45μm</span><br>'
+    assert image_mask['layout']['uirevision']
+
+
 def test_basic_return_local_file_dialog():
     assert isinstance(add_local_file_dialog(use_local_dialog=True), dbc.Button)
     assert isinstance(add_local_file_dialog(use_local_dialog=False), html.Div)